workspaces/Visual_inspection/Gauge_reading/VLM_based_gague_reading_method/data_machine/requirements.txt
--extra-index-url https://download.blender.org/pypi/
bpy==3.6.0
matplotlib
mathutils
pillow
numpy >= 2.0, <= 2.3
loguru
opencv-python <= 4.8.0.76
